requirements.txt
pyzmq
PySDL2
pyserial

